Reduce the number of loops to prevent array index out of bounds
In function sncndn, the number of loops exceeds the size of the array em, en. In some cases it returns incorrect results due to the program accessing unallocated memory.

How to fix?
This can be fixed by reducing the number of loops in function sncndn to prevent array index out of bounds.